refactor(sync-service): split 893-line file under 500-LOC limit
User mandated no file > 500 LOC. sync-service.ts had a 400-line
sync() method plus 9 private helpers. Extracted helpers into focused
modules under sync/ — orchestrator went from 893 → 489 LOC.

Modules created:
  - sync/defaults.ts      (60)  emptyGitData/Stats/Commands/Stack
  - sync/state-update.ts  (118) ensureProjectDirectories,
                                updateProjectDoc, updateStateDoc,
                                logSyncEvent
  - sync/persistence.ts   (232) recordSyncMetrics, saveDraftAnalysis,
                                archiveStaleData
  - sync/incremental.ts   (96)  detectIncrementalChanges

Bonus cleanup: removed `autoLearnFromHistory` (already a no-op since
the outcome subsystem was deleted).

Behavior unchanged — typecheck, biome, 1029 tests, build all green.
